feat: added CJS support to styled package
