+++ .changeset/shy-jars-greet.md
@@ -1,0 +1,8 @@
+---
+"@brifui/styled": patch
+"@brifui/theme": patch
+"@brifui/node": patch
+"@brifui/cli": patch
+---
+
+supports styled to be imported as CJS
