+++ .gitignore
@@ -34,6 +34,7 @@ coverage
 out/
 build
 dist
+distcjs
 !packages/cli/dist
 .install-state.gz
 *.buildinfo
