+++ package.json
@@ -14,6 +14,7 @@
     "build": "turbo build",
     "build:docs": "turbo build --filter=\"docs\"",
     "build:cli": "turbo build --filter=\"@brifui/cli\"",
+    "build:node": "turbo build --filter=\"@brifui/node\" ",
     "dev": "turbo dev --concurrency 30",
     "lint": "turbo lint",
     "test": "turbo run test",
