+++ packages/cli/src/index.ts
@@ -1,4 +1,1 @@
-#!/usr/bin/env node
-import { main } from "./main.js";
-
-main();
+export * from "./main";
