+++ packages/node/README.md
@@ -1,0 +1,1 @@
+# @brifui/node
