+++ packages/node/clean-package.config.json
@@ -1,0 +1,8 @@
+{
+  "indent": 2,
+  "replace": {
+    "main": "dist/index.cjs",
+    "module": "dist/index.js",
+    "types": "dist/index.d.ts"
+  }
+}
