+++ packages/node/package.json
@@ -1,0 +1,57 @@
+{
+  "name": "@brifui/node",
+  "version": "0.0.0",
+  "description": "Brif UI components - node utils library",
+  "sideEffects": false,
+  "type": "module",
+  "keywords": [
+    "node",
+    "utils",
+    "ui"
+  ],
+  "files": [
+    "dist",
+    "src"
+  ],
+  "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "tsup src --dts",
+    "dev": "tsup src --watch",
+    "prepack": "clean-package",
+    "postpack": "clean-package restore",
+    "lint": "eslint ."
+  },
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "dependencies": {
+    "escalade": "^3.2.0"
+  },
+  "peerDependencies": {
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
+  },
+  "devDependencies": {
+    "@brifui/eslint-config": "^0.0.3",
+    "@brifui/typescript-config": "^0.0.3",
+    "@microsoft/api-extractor": "^7.49.1",
+    "eslint": "^9.15.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "tsup": "^8.3.5",
+    "typescript": "5.5.4",
+    "vitest": "^2.1.8"
+  },
+  "tsup": {
+    "minify": false,
+    "clean": true,
+    "target": "es2019",
+    "format": [
+      "cjs",
+      "esm"
+    ]
+  }
+}
